Define EvaluationSink and CallbackSink once in _online.py
Previously the Protocol and `CallbackSink` wrapper were duplicated
between `_online.py` (for internal dispatch) and `online.py` (for the
public API). Now that `SinkPayload` already lives in `_online.py` and
gets re-exported, the sink types follow the same pattern — one
canonical definition in `_online.py` with the user-facing docstrings,
imported back into `online.py`.
